# Use Microsoft.Toolkit.HighPerformance