--- a/Directory.Build.props
+++ b/Directory.Build.props
@@ -40,6 +40,7 @@
     <VersionSuffix></VersionSuffix>
   </PropertyGroup>
   <ItemGroup>
+    <Using Include="Microsoft.Toolkit.HighPerformance" />
     <Using Include="System.Drawing" />
     <Using Include="System.Globalization" />
     <Using Include="System.Text" />