--- a/src/AdventOfCode/AdventOfCode.csproj
+++ b/src/AdventOfCode/AdventOfCode.csproj
@@ -14,6 +14,7 @@
     <EmbeddedResource Include="Input\**\input.txt" />
   </ItemGroup>
   <ItemGroup>
+    <PackageReference Include="Microsoft.Toolkit.HighPerformance" />
     <PackageReference Include="Microsoft.TypeScript.MSBuild" PrivateAssets="all" />
     <PackageReference Include="System.Runtime.Experimental" />
     <None Remove="scripts\ts\**\*.ts" />