--- a/src/AdventOfCode/Puzzles/Y2015/Day03.cs
+++ b/src/AdventOfCode/Puzzles/Y2015/Day03.cs
@@ -23,11 +23,10 @@ public class Day03 : Puzzle
     /// Gets the number of unique houses that Santa delivers at least one present to.
     /// </summary>
     /// <param name="instructions">The instructions Santa should follow.</param>
-    /// <param name="logger">The logger to use.</param>
     /// <returns>The number of unique houses that receive a delivery of at least one present.</returns>
-    internal static int GetUniqueHousesVisitedBySanta(string instructions, ILogger logger)
+    internal static int GetUniqueHousesVisitedBySanta(string instructions)
     {
-        var directions = GetDirections(instructions, logger);
+        var directions = GetDirections(instructions);
 
         var santa = new SantaGps();
         var coordinates = new HashSet<Point>(directions.Count);
@@ -45,11 +44,10 @@ public class Day03 : Puzzle
     /// Gets the number of unique houses that Santa and Robo-Santa deliver at least one present to.
     /// </summary>
     /// <param name="instructions">The instructions that Santa and Robo-Santa should follow.</param>
-    /// <param name="logger">The logger to use.</param>
     /// <returns>The number of unique houses that receive a delivery of at least one present.</returns>
-    internal static int GetUniqueHousesVisitedBySantaAndRoboSanta(string instructions, ILogger logger)
+    internal static int GetUniqueHousesVisitedBySantaAndRoboSanta(string instructions)
     {
-        var directions = GetDirections(instructions, logger);
+        var directions = GetDirections(instructions);
 
         var santa = new SantaGps();
         var roboSanta = new SantaGps();
@@ -77,8 +75,8 @@ public class Day03 : Puzzle
     {
         string instructions = await ReadResourceAsStringAsync();
 
-        HousesWithPresentsFromSanta = GetUniqueHousesVisitedBySanta(instructions, Logger);
-        HousesWithPresentsFromSantaAndRoboSanta = GetUniqueHousesVisitedBySantaAndRoboSanta(instructions, Logger);
+        HousesWithPresentsFromSanta = GetUniqueHousesVisitedBySanta(instructions);
+        HousesWithPresentsFromSantaAndRoboSanta = GetUniqueHousesVisitedBySantaAndRoboSanta(instructions);
 
         if (Verbose)
         {
@@ -98,38 +96,21 @@ public class Day03 : Puzzle
     /// Reads the directions from the specified file.
     /// </summary>
     /// <param name="instructions">The path of the file containing the directions.</param>
-    /// <param name="logger">The logger to use.</param>
     /// <returns>An <see cref="List{T}"/> containing the directions from from the specified file.</returns>
-    private static List<CardinalDirection> GetDirections(string instructions, ILogger logger)
+    private static List<CardinalDirection> GetDirections(string instructions)
     {
         var directions = new List<CardinalDirection>(instructions.Length);
 
-        foreach (var item in instructions.Enumerate())
+        for (int i = 0; i < instructions.Length; i++)
         {
-            CardinalDirection direction;
-
-            switch (item.Value)
+            CardinalDirection direction = instructions[i] switch
             {
-                case '^':
-                    direction = CardinalDirection.North;
-                    break;
-
-                case 'v':
-                    direction = CardinalDirection.South;
-                    break;
-
-                case '>':
-                    direction = CardinalDirection.East;
-                    break;
-
-                case '<':
-                    direction = CardinalDirection.West;
-                    break;
-
-                default:
-                    logger.WriteLine("Invalid direction: '{0}'.", item.Value);
-                    continue;
-            }
+                '^' => CardinalDirection.North,
+                'v' => CardinalDirection.South,
+                '>' => CardinalDirection.East,
+                '<' => CardinalDirection.West,
+                _ => throw new PuzzleException($"Invalid direction: '{instructions[i]}'."),
+            };
 
             directions.Add(direction);
         }