--- a/src/AdventOfCode/Puzzles/Y2015/Day09.cs
+++ b/src/AdventOfCode/Puzzles/Y2015/Day09.cs
@@ -124,7 +124,7 @@ public sealed class Day09 : Puzzle
 
                 visited.Remove(location);
 
-                return new() { distance };
+                return new(1) { distance };
             }
 
             var distances = new List<int>();