… src/AdventOfCode/ArrayExtensions.cs
… src/AdventOfCode/Puzzles/Y2015/Day11.cs
… src/AdventOfCode/Puzzles/Y2015/Day13.cs
… src/AdventOfCode/Puzzles/Y2015/Day14.cs
… src/AdventOfCode/Puzzles/Y2015/Day15.cs
… src/AdventOfCode/Puzzles/Y2015/Day18.cs
… src/AdventOfCode/Puzzles/Y2015/Day19.cs
… src/AdventOfCode/Puzzles/Y2015/Day20.cs
… src/AdventOfCode/Puzzles/Y2015/Day21.cs
… src/AdventOfCode/Puzzles/Y2015/Day22.cs
… src/AdventOfCode/Puzzles/Y2016/Day01.cs
… src/AdventOfCode/Puzzles/Y2016/Day02.cs
… src/AdventOfCode/Puzzles/Y2016/Day03.cs
… src/AdventOfCode/Puzzles/Y2016/Day04.cs
… src/AdventOfCode/Puzzles/Y2016/Day07.cs
… src/AdventOfCode/Puzzles/Y2016/Day08.cs
… src/AdventOfCode/Puzzles/Y2016/Day09.cs
… src/AdventOfCode/Puzzles/Y2016/Day10.cs
… src/AdventOfCode/Puzzles/Y2016/Day11.cs
… src/AdventOfCode/Puzzles/Y2016/Day12.cs
… src/AdventOfCode/Puzzles/Y2016/Day13.cs
… src/AdventOfCode/Puzzles/Y2016/Day14.cs
… src/AdventOfCode/Puzzles/Y2016/Day15.cs
… src/AdventOfCode/Puzzles/Y2016/Day16.cs
… src/AdventOfCode/Puzzles/Y2016/Day17.cs
… src/AdventOfCode/Puzzles/Y2016/Day18.cs
… src/AdventOfCode/Puzzles/Y2016/Day20.cs
… src/AdventOfCode/Puzzles/Y2016/Day21.cs
… src/AdventOfCode/Puzzles/Y2016/Day22.cs
… src/AdventOfCode/Puzzles/Y2016/Day23.cs