Add the `Microsoft.Toolkit.HighPerformance` NuGet package and use various methods it provides to either improve performance or reduce the boilerplate required to solve various puzzles.

Also various other improvements uncovered while reviewing all the puzzle implementations.