@martincostello 2022-01-09
### Baseline (8aad5d753ca49e46292217d8bce901cd1fd185c5)

``` ini

BenchmarkDotNet=v0.13.1, OS=Windows 10.0.19043.1415 (21H1/May2021Update)
Intel Core i9-9980HK CPU 2.40GHz, 1 CPU, 16 logical and 8 physical cores
.NET SDK=6.0.101
  [Host]     : .NET 6.0.1 (6.0.121.56705), X64 RyuJIT
  DefaultJob : .NET 6.0.1 (6.0.121.56705), X64 RyuJIT


```
|    Method |                input |             Mean |          Error |         StdDev |       Gen 0 |      Gen 1 |     Gen 2 |    Allocated |
|---------- |--------------------- |-----------------:|---------------:|---------------:|------------:|-----------:|----------:|-------------:|
| **Solve2015** |                   **01** |        **34.787 μs** |      **0.6658 μs** |      **0.5560 μs** |      **4.3335** |     **0.2441** |         **-** |        **36 KB** |
| Solve2016 |                   01 |       209.599 μs |      2.4941 μs |      2.3330 μs |      6.1035 |          - |         - |        51 KB |
| Solve2017 |                   01 |        13.900 μs |      0.1780 μs |      0.1390 μs |      2.6703 |     0.0305 |         - |        22 KB |
| Solve2019 |                   01 |        17.321 μs |      0.3057 μs |      0.2860 μs |      2.1057 |          - |         - |        17 KB |
| Solve2020 |                   01 |   327,014.160 μs |  3,786.6591 μs |  3,542.0432 μs |  75000.0000 |          - |         - |   616,982 KB |
| Solve2021 |                   01 |       160.137 μs |      1.6405 μs |      1.5345 μs |     28.5645 |     1.2207 |         - |       235 KB |
| **Solve2015** |                   **02** |       **263.087 μs** |      **3.2270 μs** |      **3.0186 μs** |     **54.1992** |    **10.2539** |         **-** |       **446 KB** |
| Solve2016 |                   02 |       152.877 μs |      2.4816 μs |      2.1999 μs |      4.1504 |          - |         - |        35 KB |
| Solve2017 |                   02 |        31.907 μs |      0.3507 μs |      0.3280 μs |      3.2959 |     0.0610 |         - |        27 KB |
| Solve2018 |                   02 |     1,701.631 μs |     33.3636 μs |     35.6987 μs |     93.7500 |     7.8125 |         - |       775 KB |
| Solve2019 |                   02 |        14.126 μs |      0.1899 μs |      0.1683 μs |      2.5330 |     0.0305 |         - |        21 KB |
| Solve2020 |                   02 |       546.985 μs |      5.1406 μs |      4.5570 μs |     97.6563 |    19.5313 |         - |       798 KB |
| Solve2021 |                   02 |       211.736 μs |      0.6648 μs |      0.6219 μs |     39.7949 |     5.1270 |         - |       325 KB |
| **Solve2015** |                   **03** |       **971.369 μs** |      **3.6630 μs** |      **3.0588 μs** |     **48.8281** |    **15.6250** |         **-** |       **403 KB** |
| Solve2016 |                   03 |       781.529 μs |     14.7146 μs |     15.7444 μs |    162.1094 |    53.7109 |         - |     1,325 KB |
| Solve2019 |                   03 |   145,131.193 μs |  2,341.5608 μs |  2,190.2974 μs |   2000.0000 |  1500.0000 | 1000.0000 |    41,500 KB |
| Solve2020 |                   03 |       183.981 μs |      1.4367 μs |      1.2736 μs |     13.6719 |     1.4648 |         - |       113 KB |
| Solve2021 |                   03 |       322.933 μs |      6.2562 μs |      6.4246 μs |     19.5313 |     2.9297 |         - |       161 KB |
| **Solve2017** |            **03-312051** |        **18.164 μs** |      **0.0965 μs** |      **0.0903 μs** |      **0.5798** |          **-** |         **-** |         **5 KB** |
| Solve2018 |            03-312051 |   358,005.244 μs |  6,942.0870 μs | 10,807.9930 μs |  20000.0000 | 11000.0000 | 4000.0000 |   162,425 KB |
| **Solve2016** |                   **04** |     **4,095.606 μs** |     **25.6118 μs** |     **23.9573 μs** |    **601.5625** |   **148.4375** |         **-** |     **4,918 KB** |
| Solve2017 |                   04 |     1,417.078 μs |     14.8277 μs |     13.1443 μs |    279.2969 |    46.8750 |         - |     2,294 KB |
| Solve2018 |                   04 |     1,399.937 μs |      8.8398 μs |      8.2688 μs |    130.8594 |     1.9531 |         - |     1,071 KB |
| Solve2020 |                   04 |     1,899.508 μs |     13.3196 μs |     12.4592 μs |    136.7188 |    42.9688 |         - |     1,130 KB |
| Solve2021 |                   04 |       876.544 μs |      9.6793 μs |      8.5804 μs |     46.8750 |    11.7188 |         - |       389 KB |
| **Solve2019** |     **04-138241-674034** |    **17,969.074 μs** |    **136.5795 μs** |    **127.7566 μs** |   **5156.2500** |          **-** |         **-** |    **42,126 KB** |
| **Solve2015** |        **04-iwrupvqb-5** |   **111,100.840 μs** |  **3,424.4690 μs** |  **9,935.0055 μs** |  **23333.3333** |  **8166.6667** | **8166.6667** |   **162,397 KB** |
| **Solve2017** |                   **05** |   **134,697.813 μs** |  **1,568.0056 μs** |  **1,466.7134 μs** |           **-** |          **-** |         **-** |       **141 KB** |
| Solve2020 |                   05 |       164.000 μs |      2.7208 μs |      3.1332 μs |     14.8926 |     1.9531 |         - |       123 KB |
| Solve2021 |                   05 |    26,515.657 μs |     68.0639 μs |     60.3369 μs |   1656.2500 |  1562.5000 | 1531.2500 |    17,647 KB |
| **Solve2015** |                 **05-1** |       **324.595 μs** |      **1.8588 μs** |      **1.6478 μs** |     **33.6914** |     **6.3477** |         **-** |       **276 KB** |
| Solve2019 |                 05-1 |        36.157 μs |      0.2633 μs |      0.2463 μs |      7.7515 |     0.4883 |         - |        64 KB |
| **Solve2016** |                   **06** |       **317.009 μs** |      **1.5541 μs** |      **1.3777 μs** |     **29.2969** |     **3.4180** |         **-** |       **239 KB** |
| Solve2017 |                   06 |     2,652.072 μs |     41.0681 μs |     38.4151 μs |    199.2188 |    66.4063 |         - |     1,637 KB |
| Solve2018 |                   06 |   598,690.507 μs |  5,426.0506 μs |  5,075.5311 μs |  74000.0000 |  2000.0000 | 1000.0000 |   616,663 KB |
| Solve2020 |                   06 |       933.120 μs |     11.3323 μs |      9.4629 μs |    125.9766 |    48.8281 |         - |     1,031 KB |
| Solve2021 |                   06 |       121.398 μs |      0.6770 μs |      0.6001 μs |      2.4414 |          - |         - |        21 KB |
| **Solve2015** |                 **06-1** | **1,834,248.254 μs** | **20,479.5076 μs** | **17,101.3163 μs** |   **1000.0000** |  **1000.0000** | **1000.0000** |    **27,461 KB** |
| **Solve2015** |                   **07** |     **5,041.514 μs** |     **15.5583 μs** |     **12.9919 μs** |     **31.2500** |          **-** |         **-** |       **310 KB** |
| Solve2016 |                   07 |     5,426.461 μs |     21.0104 μs |     17.5447 μs |    710.9375 |   351.5625 |         - |     5,841 KB |
| Solve2017 |                   07 |     1,644.563 μs |      9.4761 μs |      8.8640 μs |    275.3906 |   136.7188 |         - |     2,254 KB |
| Solve2018 |                   07 |     8,571.464 μs |     57.9893 μs |     51.4060 μs |    312.5000 |    15.6250 |         - |     2,584 KB |
| Solve2019 |                   07 |     3,411.567 μs |     19.4249 μs |     15.1657 μs |   1472.6563 |   140.6250 |         - |    12,060 KB |
| Solve2021 |                   07 |    36,254.317 μs |    261.3164 μs |    231.6502 μs |           - |          - |         - |       245 KB |
| **Solve2020** |        **07-shiny gold** |    **22,498.696 μs** |    **421.3355 μs** |    **394.1175 μs** |    **218.7500** |    **93.7500** |         **-** |     **1,930 KB** |
| **Solve2015** |                   **08** |       **150.344 μs** |      **0.8165 μs** |      **0.7638 μs** |     **28.8086** |     **2.1973** |         **-** |       **236 KB** |
| Solve2016 |                   08 |       132.610 μs |      0.6833 μs |      0.6057 μs |     19.2871 |     1.2207 |         - |       159 KB |
| Solve2017 |                   08 |     1,289.182 μs |     13.4527 μs |     11.2336 μs |    123.0469 |    37.1094 |         - |     1,014 KB |
| Solve2018 |                   08 |     1,639.448 μs |     20.6042 μs |     17.2055 μs |    181.6406 |    99.6094 |   41.0156 |     1,415 KB |
| Solve2019 |                   08 |       221.000 μs |      4.2681 μs |      4.5668 μs |     20.2637 |     4.8828 |         - |       167 KB |
| Solve2020 |                   08 |     1,150.063 μs |      8.8910 μs |      8.3166 μs |    363.2813 |    58.5938 |         - |     2,979 KB |
| Solve2021 |                   08 |     3,013.714 μs |     29.9749 μs |     26.5720 μs |    332.0313 |   109.3750 |         - |     2,728 KB |
| **Solve2015** |                   **09** |    **42,218.786 μs** |    **410.2842 μs** |    **363.7063 μs** |   **3416.6667** |   **583.3333** |         **-** |    **28,565 KB** |
| Solve2016 |                   09 |       229.738 μs |      0.9797 μs |      0.8685 μs |     52.9785 |     7.3242 |         - |       433 KB |
| Solve2017 |                   09 |       101.330 μs |      1.9333 μs |      1.8084 μs |     14.5264 |     2.8076 |         - |       120 KB |
| Solve2020 |                   09 |     3,347.369 μs |     29.7519 μs |     26.3743 μs |    511.7188 |    27.3438 |         - |     4,191 KB |
| Solve2021 |                   09 |     5,451.859 μs |     16.0437 μs |     13.3973 μs |    390.6250 |   195.3125 |  195.3125 |     3,100 KB |
| **Solve2019** |                 **09-1** |        **53.667 μs** |      **1.0453 μs** |      **1.0266 μs** |     **12.9395** |     **1.4038** |         **-** |       **106 KB** |
| **Solve2016** |                   **10** |       **127.853 μs** |      **1.0870 μs** |      **1.0168 μs** |     **31.7383** |     **5.1270** |         **-** |       **261 KB** |
| Solve2017 |                   10 |     1,283.482 μs |      6.4844 μs |      5.7483 μs |     97.6563 |          - |         - |       803 KB |
| Solve2020 |                   10 |       196.219 μs |      2.0153 μs |      1.7865 μs |     10.9863 |     0.7324 |         - |        91 KB |
| Solve2021 |                   10 |       137.277 μs |      0.6884 μs |      0.5749 μs |      6.3477 |     0.4883 |         - |        53 KB |
| **Solve2017** |                   **11** |   **246,968.215 μs** |  **1,808.1855 μs** |  **1,509.9167 μs** |           **-** |          **-** |         **-** |       **888 KB** |
| Solve2020 |                   11 |    98,703.910 μs |    448.3974 μs |    374.4321 μs |    833.3333 |   166.6667 |         - |     8,115 KB |
| Solve2021 |                   11 |     7,435.902 μs |     18.0570 μs |     16.8905 μs |    320.3125 |     7.8125 |         - |     2,645 KB |
| **Solve2015** |          **11-cqjxjnds** |     **3,383.148 μs** |      **9.1358 μs** |      **8.0986 μs** |      **3.9063** |          **-** |         **-** |        **46 KB** |
| **Solve2015** |                   **12** |       **722.409 μs** |      **1.5894 μs** |      **1.4089 μs** |      **4.8828** |          **-** |         **-** |        **44 KB** |
| Solve2017 |                   12 |   107,357.533 μs |    414.1225 μs |    367.1088 μs |   2400.0000 |   800.0000 |  400.0000 |    20,333 KB |
| Solve2020 |                   12 |       139.637 μs |      1.3468 μs |      1.2598 μs |     21.9727 |     2.1973 |         - |       180 KB |
| Solve2021 |                   12 |   101,678.341 μs |    923.2566 μs |    863.6148 μs |   2400.0000 |          - |         - |    19,744 KB |
| **Solve2017** |                   **13** |    **23,841.882 μs** |    **126.1760 μs** |    **118.0251 μs** |           **-** |          **-** |         **-** |        **27 KB** |
| Solve2020 |                   13 |        50.462 μs |      0.9887 μs |      0.9248 μs |      4.8828 |          - |         - |        40 KB |
| Solve2021 |                   13 |       661.901 μs |      4.6105 μs |      4.0871 μs |     72.2656 |    11.7188 |         - |       591 KB |
| **Solve2016** |              **13-1362** |    **27,199.959 μs** |     **97.0698 μs** |     **90.7991 μs** |   **1781.2500** |    **62.5000** |         **-** |    **14,556 KB** |
| **Solve2020** |                   **14** |     **9,134.068 μs** |     **27.7094 μs** |     **23.1386 μs** |   **1031.2500** |   **968.7500** |  **968.7500** |     **6,075 KB** |
| Solve2021 |                   14 |       633.872 μs |      6.2392 μs |      5.8362 μs |     72.2656 |     3.9063 |         - |       592 KB |
| **Solve2015** |              **14-2503** |     **1,717.181 μs** |      **3.5178 μs** |      **3.1185 μs** |    **234.3750** |     **1.9531** |         **-** |     **1,918 KB** |
| **Solve2017** |          **14-hwlqcszp** |    **69,285.552 μs** |    **270.0609 μs** |    **252.6151 μs** |   **5500.0000** |          **-** |         **-** |    **45,294 KB** |
| **Solve2016** |                   **15** |   **138,719.357 μs** |    **831.5730 μs** |    **777.8539 μs** |  **44250.0000** |          **-** |         **-** |   **361,663 KB** |
| Solve2021 |                   15 |   229,174.029 μs |  1,452.7767 μs |  1,287.8487 μs |   4333.3333 |  1666.6667 | 1333.3333 |    62,168 KB |
| **Solve2015** |                   **16** |       **979.289 μs** |      **7.2343 μs** |      **6.7670 μs** |    **179.6875** |    **29.2969** |         **-** |     **1,478 KB** |
| Solve2020 |                   16 |    16,396.065 μs |    227.0476 μs |    201.2718 μs |   5593.7500 |   281.2500 |   31.2500 |    45,804 KB |
| Solve2021 |                   16 |        81.340 μs |      0.6182 μs |      0.5783 μs |     24.2920 |     0.8545 |         - |       199 KB |
| **Solve2016** | **16-10(...)0-272 [24]** |         **7.190 μs** |      **0.1394 μs** |      **0.1605 μs** |      **0.9460** |          **-** |         **-** |         **8 KB** |
| **Solve2021** |                   **17** |    **10,460.094 μs** |     **64.4862 μs** |     **60.3204 μs** |   **2703.1250** |   **140.6250** |   **15.6250** |    **22,170 KB** |
| **Solve2015** |               **17-150** |    **42,845.768 μs** |    **211.0317 μs** |    **197.3991 μs** |    **500.0000** |   **500.0000** |  **500.0000** |     **8,404 KB** |
| **Solve2016** |          **17-pvhmgsws** |    **64,127.753 μs** |    **278.2715 μs** |    **260.2953 μs** |  **12500.0000** |   **500.0000** |  **125.0000** |   **102,635 KB** |
| **Solve2020** |                   **18** |       **964.616 μs** |      **6.0205 μs** |      **5.3370 μs** |    **192.3828** |    **31.2500** |         **-** |     **1,572 KB** |
| Solve2021 |                   18 |   202,069.064 μs |  1,865.2490 μs |  1,744.7551 μs |  30000.0000 |          - |         - |   245,526 KB |
| **Solve2015** |         **18-100-false** |   **111,010.052 μs** |    **672.5714 μs** |    **629.1237 μs** |  **14400.0000** |  **1000.0000** |         **-** |   **118,270 KB** |
| **Solve2016** |                **18-40** |        **46.602 μs** |      **0.2238 μs** |      **0.2094 μs** |      **1.2207** |          **-** |         **-** |        **10 KB** |
| **Solve2020** |                   **19** |     **6,083.160 μs** |     **67.6966 μs** |     **63.3234 μs** |     **70.3125** |    **15.6250** |         **-** |       **587 KB** |
| Solve2021 |                   19 | 1,574,752.593 μs | 14,029.0314 μs | 13,122.7647 μs | 181000.0000 |  1000.0000 |         - | 1,485,870 KB |
| **Solve2016** |         **19-3014387-1** |   **298,078.491 μs** |  **4,850.6362 μs** |  **4,763.9733 μs** |  **18500.0000** | **10000.0000** | **2000.0000** |   **141,304 KB** |
| **Solve2016** |         **19-3014387-2** |   **303,112.936 μs** |  **5,796.0732 μs** |  **7,118.0985 μs** |  **18500.0000** | **10000.0000** | **2000.0000** |   **141,303 KB** |
| **Solve2015** |         **19-calibrate** |     **4,651.520 μs** |     **25.5635 μs** |     **21.3467 μs** |   **2742.1875** |   **257.8125** |  **109.3750** |    **22,435 KB** |
| **Solve2016** |                   **20** |     **8,062.882 μs** |     **34.7381 μs** |     **32.4941 μs** |    **640.6250** |   **140.6250** |         **-** |     **5,293 KB** |
| Solve2020 |                   20 |     8,144.635 μs |     74.8594 μs |     70.0235 μs |    468.7500 |   203.1250 |         - |     3,876 KB |
| Solve2021 |                   20 |   510,752.600 μs |  1,889.0181 μs |  1,766.9887 μs |  15000.0000 | 10000.0000 | 9000.0000 |   105,006 KB |
| **Solve2015** |                   **21** |       **645.546 μs** |      **3.2209 μs** |      **2.6896 μs** |    **199.2188** |     **6.8359** |         **-** |     **1,633 KB** |
| Solve2020 |                   21 |       420.919 μs |      2.2108 μs |      1.8461 μs |     31.2500 |     8.3008 |         - |       257 KB |
| Solve2021 |                   21 |   219,032.411 μs |  3,256.6411 μs |  3,046.2641 μs |  45333.3333 |  2000.0000 | 1666.6667 |   390,339 KB |
| **Solve2016** |          **21-abcdefgh** |        **35.256 μs** |      **0.1293 μs** |      **0.1079 μs** |      **6.7139** |     **0.2441** |         **-** |        **55 KB** |
| **Solve2016** |                   **22** |     **7,971.051 μs** |     **57.0395 μs** |     **53.3547 μs** |     **93.7500** |    **31.2500** |         **-** |       **883 KB** |
| Solve2020 |                   22 |   147,172.712 μs |  2,818.6734 μs |  2,636.5889 μs |  20750.0000 |  1750.0000 |  250.0000 |   170,460 KB |
| Solve2021 |                   22 |   204,363.189 μs |  4,036.7776 μs |  5,248.9512 μs |   6333.3333 |  5666.6667 | 5333.3333 |   143,468 KB |
| **Solve2015** |                   **23** |       **173.575 μs** |      **1.1497 μs** |      **0.8976 μs** |     **17.3340** |          **-** |         **-** |       **143 KB** |
| Solve2021 |                   23 |   825,113.500 μs | 14,672.4205 μs | 15,067.4956 μs |  85000.0000 | 28000.0000 | 6000.0000 |   694,677 KB |
| **Solve2016** |                 **23-7** |     **5,328.811 μs** |     **31.6123 μs** |     **28.0235 μs** |    **859.3750** |          **-** |         **-** |     **7,081 KB** |
| **Solve2016** |                   **24** |    **58,054.777 μs** |  **1,153.1314 μs** |  **1,184.1810 μs** |   **5777.7778** |  **1222.2222** |  **888.8889** |    **45,922 KB** |
| Solve2020 |                   24 |   177,262.187 μs |  2,167.5177 μs |  1,809.9754 μs |   7666.6667 |  3666.6667 | 3333.3333 |    55,370 KB |
| Solve2021 |                   24 |        28.710 μs |      0.3846 μs |      0.3598 μs |      5.2795 |     0.2747 |         - |        43 KB |
| **Solve2020** |                   **25** |     **3,277.844 μs** |     **12.1487 μs** |     **10.7695 μs** |           **-** |          **-** |         **-** |         **6 KB** |
| Solve2021 |                   25 |    73,269.195 μs |    631.2239 μs |    590.4472 μs |   4571.4286 |   571.4286 |         - |    37,891 KB |
| **Solve2015** |         **25-2947-3029** |    **60,738.110 μs** |    **243.9528 μs** |    **216.2578 μs** |           **-** |          **-** |         **-** |         **1 KB** |

### This PR (e53820d8f88f7a952e201dab5ed8366826cece32)

``` ini

BenchmarkDotNet=v0.13.1, OS=Windows 10.0.19043.1415 (21H1/May2021Update)
Intel Core i9-9980HK CPU 2.40GHz, 1 CPU, 16 logical and 8 physical cores
.NET SDK=6.0.101
  [Host]     : .NET 6.0.1 (6.0.121.56705), X64 RyuJIT
  DefaultJob : .NET 6.0.1 (6.0.121.56705), X64 RyuJIT


```
|    Method |                input |             Mean |          Error |         StdDev |           Median |       Gen 0 |      Gen 1 |     Gen 2 |    Allocated |
|---------- |--------------------- |-----------------:|---------------:|---------------:|-----------------:|------------:|-----------:|----------:|-------------:|
| **Solve2015** |                   **01** |        **32.256 μs** |      **0.1505 μs** |      **0.1334 μs** |        **32.230 μs** |      **4.3335** |     **0.2441** |         **-** |        **36 KB** |
…